refactor: use always `convert.base64-decode` stream filter

+++ src/Controller/StreamsController.php
@@ -50,32 +50,15 @@ class StreamsController extends ResourcesController
     public function beforeFilter(Event $event)
     {
         // Decode base64-encoded body.
-        if ($this->request->getHeaderLine('Content-Transfer-Encoding') === 'base64') {
-            // Check if any suitable stream filter is available.
-            $filter = null;
-            $filters = stream_get_filters();
-            if (in_array('string.base64', $filters)) {
-                $filter = 'string.base64';
-            } elseif (in_array('convert.*', $filters)) {
-                $filter = 'convert.base64-decode';
-            }
+        if ($this->request->getParam('action') === 'upload' && $this->request->getHeaderLine('Content-Transfer-Encoding') === 'base64') {
+            // Append filter to stream.
+            $body = $this->request->getBody();
 
-            if (!empty($filter)) {
-                // Append filter to stream.
-                $body = $this->request->getBody();
+            $stream = $body->detach();
+            stream_filter_append($stream, 'convert.base64-decode', STREAM_FILTER_READ);
 
-                $stream = $body->detach();
-                stream_filter_append($stream, $filter, STREAM_FILTER_READ);
-
-                $body = new Stream($stream, 'r');
-                $this->request = $this->request->withBody($body);
-            } else {
-                // No suitable filter available. Read whole stream and decode it.
-                $body = new Stream('php://temp', 'wb+');
-                $body->write(base64_decode($this->request->getBody()->getContents()));
-                $body->rewind();
-                $this->request = $this->request->withBody($body);
-            }
+            $body = new Stream($stream, 'r');
+            $this->request = $this->request->withBody($body);
         }
 
         return parent::beforeFilter($event);
